버그 수정
1. delete character 순서 관련 문제 해결
2. task가 끝나면 task에 쓰레기값이 들어가 nullptr로 초기화되지 않아 발생하는 문제 해결: 원본이 아닌 포인터 변수에 nullptr를 할당하면 원본에는 nullptr가 들어가지 않아 쓰레기값만 남게 됨.

+++ CPP/NativeLib/NativeLib/ObjectService.h
@@ -55,9 +55,14 @@ public:
 	}
 
 	void DeleteCharacter(int character_id) {
-		deleted_character_ids.push(character_id);
+		if (object_repo->IsNotExistId(character_id)) {
+			printf("WARNING: [ObjectService] Trying to delete not exist character. ");
+			printf("id:% d\n", character_id);
+			return;
+		}
 		factory.ReturnCharacterName(object_repo->GetCharacter(character_id)->GetFullName());
-		object_repo->DeleteCharacter(character_id);
+		deleted_character_ids.push(character_id);
+		bool result = object_repo->DeleteCharacter(character_id);
 	}
 
 	bool IsCharacterNotExist(int character_id) {
